# implementation_examples/openapi.yaml
openapi: 3.1.0
...
info:
  title: Oolix API
  version: 0.2.0
  description: |
    Privacy-safe partner media activation control plane.

    **The rule every endpoint serves.** Oolix is the control plane; Data
    Partners remain the customer-data owners (§1). No request or response in
    this contract carries a `partner_user_id`, a segment membership list or an
    exact audience count. Reach is always a bucket (§72) and attribution is
    always an opaque token resolved server-side by SHA-256 (§90).
  contact:
    name: Oolix engineering
  license:
    name: Oolix proprietary
    identifier: LicenseRef-Oolix-Proprietary
servers:
  - url: http://localhost:4000
    description: Local development (docker compose stack)
  - url: '{scheme}://{host}'
    description: |
      Deployed environments. The host is supplied per environment rather than
      hard-coded, so a generated client can never default to someone else's
      domain.
    variables:
      scheme: { default: https, enum: [https] }
      host: { default: localhost:4000 }
...
security:
  - oidcBearer: []
paths:
...
  /v1/me/context:
    get:
      operationId: getMyContext
      parameters:
        - $ref: '#/components/parameters/OrgId'
      tags: [Identity]
      summary: Resolve the caller's organizations, roles and permissions
      description: |
        §52.1. Everything the portal needs to render role-aware navigation
        (§34) in one call, so the frontend never infers permissions from a role
        name. Roles come from the database, never from a token claim (§4.2).
      responses:
        '200':
          description: Caller context
        '401': { $ref: '#/components/responses/Unauthorized' }

        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/organizations:
    post:
      operationId: createOrganization
      parameters:
        - $ref: '#/components/parameters/OrgId'
      tags: [Identity]
      summary: Create an organization
      description: |
        §35.2. New organizations start at `BUSINESS_VERIFICATION_PENDING`;
        §66.3 permits browsing and drafting while pending, but not submitting
        or publishing.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateOrganization' }
      responses:
        '201': { description: Organization created }
        '400': { $ref: '#/components/responses/ValidationError' }

        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/organizations/members:
    get:
      operationId: listOrganizationMembers
      parameters:
        - $ref: '#/components/parameters/OrgId'
      tags: [Identity]
      summary: List organization members
      responses:
        '200': { description: Members }
        '403': { $ref: '#/components/responses/Forbidden' }

        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/organizations/members/invite:
    post:
      operationId: inviteOrganizationMember
      parameters:
        - $ref: '#/components/parameters/OrgId'
      tags: [Identity]
      summary: Invite a member
      responses:
        '201': { description: Invited }
        '403': { $ref: '#/components/responses/Forbidden' }

        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/organizations/members/{userId}:
    delete:
      operationId: removeOrganizationMember
      tags: [Identity]
      summary: Remove a member
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - { name: userId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Removed }
        '403': { $ref: '#/components/responses/Forbidden' }

        '429': { $ref: '#/components/responses/RateLimited' }
  # ---------------------------------------------------------------------------
...
  /v1/campaigns:
    get:
      operationId: listCampaigns
      parameters:
        - $ref: '#/components/parameters/OrgId'
      tags: [Campaign]
      summary: List campaigns
      responses:
        '200': { description: Campaigns }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
    post:
      operationId: createCampaign
      tags: [Campaign]
      summary: Create a campaign draft
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCampaign' }
      responses:
        '201':
          description: Campaign created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Campaign' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '409': { $ref: '#/components/responses/IdempotencyConflict' }

        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/campaigns/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      operationId: getCampaign
      parameters:
        - $ref: '#/components/parameters/OrgId'
      tags: [Campaign]
      summary: Campaign detail with per-Partner activations
      description: |
        §42: the parent campaign is an aggregate **view**. Each Partner request
        and each activation carries its own independent state and reason;
        `derived_status` is a projection over the activations.
      responses:
        '200': { description: Campaign detail }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
    patch:
      operationId: updateCampaign
      parameters:
        - $ref: '#/components/parameters/OrgId'
      tags: [Campaign]
      summary: Update a draft campaign
      responses:
        '200': { description: Updated }
        '409': { description: 'Campaign is not DRAFT (§76)' }

        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/campaigns/{id}/audience-link:
    post:
      operationId: linkCampaignAudience
      tags: [Campaign, Audience]
      summary: Link an Audience Group to the campaign
      description: |
        v6 §9 step 3. Freezes the Audience Group version and its rule hash onto
        the campaign: §10 binds a Partner's approval to that hash, so it must
        not move when the Buyer edits the audience afterwards.

        One campaign links one Audience Group (§9). Re-linking is allowed while
        the campaign is DRAFT and no Partner request is in review.
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [audience_group_id]
              properties:
                audience_group_id: { type: string, format: uuid }
                audience_version:
                  type: integer
                  minimum: 1
                  description: Defaults to the audience's current version, frozen at link time.
      responses:
        '200': { description: Audience linked and frozen }
        '409':
          description: |
            `CAMP_002` — the campaign is not DRAFT, the audience version is not
            READY, or a Partner request is already in review.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429': { $ref: '#/components/responses/RateLimited' }
    get:
      operationId: getCampaignAudience
      tags: [Campaign, Audience]
      summary: Read the campaign's frozen audience
      description: 'v6 §9 step 10: what the review screen shows. `audience` is null when none is linked.'
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: 'The linked audience, or null when none is set' }
        '429': { $ref: '#/components/responses/RateLimited' }
    delete:
      operationId: unlinkCampaignAudience
      tags: [Campaign, Audience]
      summary: Unlink the campaign's audience
      description: |
        v6 §9 keeps step 3 revisable until submit. Refused once Partner requests
        on this campaign target the audience — removing it would leave them
        pointing at nothing.
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Audience unlinked }
        '409':
          description: '`CAMP_002` — Partner requests still target this audience.'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/campaigns/{id}/partner-requests:
    post:
      operationId: addPartnerRequest
      tags: [Campaign]
      summary: Add a Partner request
      description: |
        §40.4: one parent campaign may hold 1..N Partner requests, and every
        Partner receives only its own. §40.5: each requested channel becomes a
        separate activation -- never one ambiguous activation spanning several.
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePartnerRequest' }
      responses:
        '201': { description: Partner request created }
        '409':
          description: |
            `CAMP_003` budget exceeded (§40.6), `PART_002` Partner not ready or
            category blocked (§37, §41), or `CHAN_001` external channel not
            enabled (§84).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

        '429': { $ref: '#/components/responses/RateLimited' }
...
  /v1/activations/{id}/pause:
    post:
      operationId: pauseActivation
      tags: [Activation]
      summary: Pause an activation
      description: |
        §52.3. Reachable by the Partner that serves it **and** the Buyer that
        pays for it; neither can touch the other's activations.
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '201': { description: Paused }
        '404': { $ref: '#/components/responses/NotFound' }

        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/activations/{id}/resume:
    post:
      operationId: resumeActivation
      tags: [Activation]
      summary: Resume a paused activation
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '201': { description: Live }
        '409': { description: 'An active kill switch prevents resuming (§24)' }

        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/activations/{id}/end:
    post:
      operationId: endActivation
      tags: [Activation]
      summary: End an activation
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '201': { description: Ending }
        '400': { description: 'A reason is required for the audit trail (§83)' }

        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/partner/kill-switches:
    get:
      operationId: listKillSwitches
      parameters:
        - $ref: '#/components/parameters/OrgId'
      tags: [Activation]
      summary: List kill switches
      responses:
        '200': { description: Kill switches }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
    post:
      operationId: activateKillSwitch
      parameters:
        - $ref: '#/components/parameters/OrgId'
      tags: [Activation]
      summary: Activate a kill switch
      description: |
        §24, §6. Unilateral and immediate: no Buyer agreement, no Oolix
        approval, no notice period. Enforced **locally** by the Agent so it
        works while Oolix is unreachable. §78.2 alerts on every activation.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/KillSwitch' }
      responses:
        '201': { description: Kill switch active }

        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/partner/kill-switches/{id}/release:
    post:
      operationId: releaseKillSwitch
      tags: [Activation]
      summary: Release a kill switch
      description: |
        Releasing does **not** resurrect an ended activation: §76 has no path
        from ENDED back to LIVE.
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '201': { description: Released }

        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  # ---------------------------------------------------------------------------
...
  /v1/reports/campaigns/{id}:
    get:
      operationId: getCampaignReport
      tags: [Reporting]
      summary: Buyer campaign report
      description: |
        §49. Delivery comes from Agent aggregates; outcomes come from CRM lead
        states. They are kept distinct because they are verified differently --
        a Partner counts an impression, a **Buyer** confirms a qualified lead.

        Per-Partner figures are **not** deduplicated across Partners (§13, §72).
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Campaign report }

        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
components:
  securitySchemes:
    oidcBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        §64: OIDC access token. Identity is delegated to the provider; Oolix
        stores only the subject claim. Roles are read from the database for the
        active organization, never from a token claim (§4.2).
    agentBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: '§92.3: Oolix-issued Agent access token, 15-minute lifetime.'
    crmApiKey:
      type: http
      scheme: bearer
      description: "§71: the Buyer's CRM API key. Hashed at rest and rotatable."
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, maxLength: 200 }
      description: |
        §22.3, §99. Same key + same payload replays the stored result; same key
        + different payload is `409 IDEMPOTENCY_CONFLICT`. **Required** on
        campaign submit, approval and mark-paid (§53).
...
    OrgId:
      name: X-Org-Id
      in: header
      required: false
      schema: { type: string, format: uuid }
      description: |
        Selects which organization the request acts within, for a user who
        belongs to several (§34 role switching without a second account).
...
  responses:
    Unauthorized:
      description: '`AUTH_001` -- invalid or expired token.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: '`PERM_001` / `PERM_002` -- permission or organization scope denied.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: |
        `CAMP_001` / `PART_001`. Also returned where an object exists but is
        not visible to the caller, so existence cannot be probed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: '`VAL_001` -- schema validation failed.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: |
        `SYS_001` -- §86's per-class budget for this caller is exhausted. The
        response carries `X-RateLimit-*`, `Retry-After` and
        `retry_after_seconds`, so a client can pace itself instead of guessing
        (§94).
      headers:
        X-RateLimit-Limit:
          { schema: { type: integer }, description: Requests allowed in the window. }
        X-RateLimit-Remaining:
          { schema: { type: integer }, description: Requests left in the window. }
        X-RateLimit-Reset:
          { schema: { type: integer }, description: Unix seconds when the window frees up. }
        Retry-After: { schema: { type: integer }, description: Seconds to wait before retrying. }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    IdempotencyConflict:
      description: '`IDEMPOTENCY_CONFLICT` -- key reused with a different payload (§99).'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    Error:
      type: object
      required: [error]
      description: |
        §53, §99. Every error uses this envelope: a canonical code, a human
        message, an explicit retryable flag, field errors and the correlation
        id. §99 forbids exposing stack traces or secrets.
      properties:
        error:
          type: object
          required: [code, message, retryable, correlation_id, field_errors]
          properties:
            code: { $ref: '#/components/schemas/ErrorCode' }
            message: { type: string }
            retryable: { type: boolean }
            correlation_id: { type: string }
            retry_after_seconds: { type: integer }
            field_errors:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  message: { type: string }
    ErrorCode:
      type: string
      description: '§77.2 canonical error registry.'
      enum:
        [
          AUTH_001,
          PERM_001,
          PERM_002,
          VAL_001,
          CAMP_001,
          CAMP_002,
          CAMP_003,
          PART_001,
          PART_002,
          CHAN_001,
          CHAN_002,
          AGENT_001,
          SYS_001,
          SYS_002,
          IDEMPOTENCY_CONFLICT,
        ]

    # --- §89 canonical channel enum ------------------------------------------
    Channel:
      type: string
      description: |
        §89: exactly these four strings in the database, APIs, manifests,
        events and frontend state. Eligibility is a separate state, never
        encoded into the channel constant.
      enum: [PARTNER_WEB, PARTNER_APP, META, GOOGLE]
...
    Money:
      type: object
      description: '§53, §73: integer minor units plus ISO-4217 code. Never a float.'
      required: [amount_minor, currency]
      properties:
        amount_minor: { type: integer, format: int64, minimum: 1 }
        currency: { type: string, minLength: 3, maxLength: 3 }
    FrequencyCap:
      type: object
      required: [max_impressions, window]
      properties:
        max_impressions: { type: integer, minimum: 1, maximum: 100 }
        window:
          type: string
          description: 'ISO-8601 duration, e.g. `P1D` (§67.2).'
          pattern: '^P(?:\d+D)?(?:T(?:\d+H)?(?:\d+M)?(?:\d+S)?)?$'
...
    ActivationStatus:
      type: string
      enum: [PENDING_CHANNEL_CHECK, READY, SYNCING, LIVE, PAUSED, FAILED, ENDING, ENDED]

    # --- request bodies ------------------------------------------------------
    CreateOrganization:
      type: object
      required: [name, domain, type, country]
      properties:
        name: { type: string, minLength: 2, maxLength: 200 }
        domain: { type: string, description: 'Bare domain, verified per §66.3.' }
        type:
          { type: string, enum: [BUYER, DATA_PARTNER, BUYER_AND_PARTNER, NETWORK_SPONSOR, AGENCY] }
        country: { type: string, minLength: 2, maxLength: 2 }
        industry: { type: string }
        tax_id: { type: string }
...
    CreateCampaign:
      type: object
      required:
        [name, objective, brand_id, category, purpose_id, budget, start_at, end_at, geographies]
      properties:
        name: { type: string, minLength: 3, maxLength: 120 }
        objective: { type: string, enum: [QUALIFIED_LEADS, CONVERSIONS, CLICKS, AWARENESS] }
        brand_id: { type: string, format: uuid }
        category: { type: string }
        purpose_id:
          type: string
          description: '§81: explicit and versioned -- never a permanent "eligible" boolean.'
        budget: { $ref: '#/components/schemas/Money' }
        start_at: { type: string, format: date-time }
        end_at: { type: string, format: date-time }
        geographies: { type: array, items: { type: string }, minItems: 1 }
        landing_url:
          type: string
          format: uri
          description: |
            §40.2, §40.7: HTTPS, on the brand's verified domain, and carrying no
            PII query parameters. Required for click, lead and conversion
            objectives.
        lead_definition:
          $ref: '#/components/schemas/LeadDefinition'
    LeadDefinition:
      type: object
      description: |
        §40.8. Required for outcome objectives: §50 forbids paying a Partner on
        unverified clicks, so a CPQL campaign that never defines "qualified"
        has no settleable basis.
      required: [qualified_statuses]
      properties:
        qualified_statuses:
          type: array
          items: { type: string, enum: [VALID, QUALIFIED, CONVERTED] }
          minItems: 1
        duplicate_window_days: { type: integer, minimum: 1, maximum: 365, default: 30 }
        valid_lead_rule: { type: string }
        qualified_lead_rule: { type: string }
        conversion_event: { type: string }
    Campaign:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        objective: { type: string }
        status: { type: string }
        derived_status:
          type: string
          description: '§42: a projection over the activations, never a stored source of truth.'
        budget: { $ref: '#/components/schemas/Money' }
        version: { type: integer }
        activations:
          type: array
          description: |
            §42: each activation carries its OWN state. A campaign is routinely
            LIVE at one Partner and REJECTED at another, and flattening that
            into a single status would hide the rejection.
          items:
            type: object
            properties:
              activation_id: { type: string, format: uuid }
              partner_org_id: { type: string, format: uuid }
              channel: { $ref: '#/components/schemas/Channel' }
              status: { $ref: '#/components/schemas/ActivationStatus' }
              status_reason: { type: [string, 'null'] }
    CreatePartnerRequest:
      type: object
      required: [partner_org_id, channels, creative_version_ids]
      description: |
        v6 §19: a request targets EITHER the campaign's linked Audience Group
        (the primary path — omit `segment_id`) or a Partner's prebuilt segment
        (the legacy path). Never both.
      properties:
        partner_org_id: { type: string, format: uuid }
        segment_id:
          type: string
          format: uuid
          description: |
            v6 §19 legacy path. Omit to target the campaign's linked Audience
            Group; `409 CAMP_002` if no audience is linked either.
        reach_estimate_id:
          type: string
          format: uuid
          description: |
            v6 §9: the reach estimate the Buyer was looking at when they chose
            this Partner, frozen onto the request so the Partner approves
            against the same number. Must belong to this Partner and this exact
            audience version.
        channels:
          type: array
          minItems: 1
          items:
            type: object
            required: [channel, allocation_minor, frequency_cap]
            properties:
              channel: { $ref: '#/components/schemas/Channel' }
              placement_ids: { type: array, items: { type: string, format: uuid } }
              allocation_minor: { type: integer, minimum: 1 }
              frequency_cap: { $ref: '#/components/schemas/FrequencyCap' }
        creative_version_ids: { type: array, items: { type: string, format: uuid }, minItems: 1 }
        audience_expansion_allowed:
          type: boolean
          default: false
          description: '§40.4, §41: off unless the Partner explicitly grants it.'
        partner_payout:
          type: object
          properties:
            model: { type: string, enum: [CPM, CPC, CPL, CPQL, FIXED, HYBRID] }
            unit_price_minor: { type: integer, minimum: 0 }
...
    KillSwitch:
      type: object
      required: [scope, reason]
      properties:
        scope: { type: string, enum: [AGENT, PLACEMENT, ACTIVATION, CHANNEL, PARTNER_ALL] }
        target_id:
          type: [string, 'null']
          description: |
            For `PLACEMENT` this is the placement KEY -- what the Agent matches
            on locally without a lookup (§75 offline enforcement).
        reason: { type: string, minLength: 3, maxLength: 500 }
...